Renames attributes file
