+++ tests/unit/lbrynet_daemon/_sample_attributes_list.py
